requirements.txt
djangorestframework_simplejwt==4.4.0
Django==3.0.3
djangorestframework==3.11.0
